requirements.txt
aniso8601==9.0.1
apispec==6.4.0
apispec-webframeworks==1.0.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
Automat==22.10.0
bidict==0.23.1
binaryornot==0.4.4
blinker==1.7.0
boto3==1.34.34
botocore==1.34.79
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.5.0
deepdiff==6.7.1
dnspython==2.6.1
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
gevent==23.9.1
gevent-websocket==0.10.1
gitignore_parser==0.1.11
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
icecream==2.1.3
idna==3.6
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
kthread==0.2.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.2
ordered-set==4.1.0
packaging==24.0
pandas==2.2.0
passlib==1.7.4
plotly==5.20.0
pyarrow==15.0.0
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-slugify==8.0.4
python-socketio==5.11.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.13.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.25
taipy==3.1.1
taipy-config==3.1.1
taipy-core==3.1.1
taipy-gui==3.1.1
taipy-rest==3.1.1
taipy-templates==3.1.1
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.4.0
toml==0.10.2
Twisted==23.10.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
xgboost==2.0.3
zope.event==5.0
zope.interface==6.2